UserEvent now usize size_t for UserData fields to avoid cutting number on some platforms. Fix some warnings.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@5417 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ include/IEventReceiver.h
@@ -405,10 +405,10 @@ struct SEvent
 	struct SUserEvent
 	{
 		//! Some user specified data as int
-		s32 UserData1;
+		size_t UserData1;
 
 		//! Another user specified data as int
-		s32 UserData2;
+		size_t UserData2;
 	};
 
 	EEVENT_TYPE EventType;
